Updating i in daysBetweenDates
In the daysBetweenDates function, the for loop properly decremented the i value, but the isLeapYear() function contained the incorrect input parameter. Instead of the *year1 or *year2 variable, the input parameter needs to be i.

Will update the file later in order to stop the output of the print functions within the function toJulian().